rename time_delta_1d
